requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
spacy==3.7.2
python-multipart==0.0.6
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
transformers==4.35.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 